Use more than one 'not present' byte value to catch Decode overruns.
Previously, the dst buffer was initialized to all 0xfe byte values. An
overrun caused by an incorrect copy from an early location to a later
one (past the end of dst) could have slipped through as a false negative
because the old value and the new value were the same value (0xfe), so
no change was detected.

After this commit, the dst buffer is initialized with a variety of byte
values, which makes it less likely, although not impossible, to have
such false negatives.