Use a more inclusive text for golden input.
Replace the first chapter of Tom Sawyer with the first 400 lines of
Isaac Newton's Opticks.

The rawsnappy version was generated by cmd/snappytool in this repo.

The extendMatch test goldens were updated as per the instructions in
golden_test.go (with an update to account for the golang version of
extendMatch being inlined.)